+++ packages/table_deprecated/index.ts
@@ -1,5 +1,5 @@
-export { default as Table } from "./components/Table";
-export { default as CheckboxTable } from "./components/CheckboxTable";
+export { default as Table_Deprecated } from "./components/Table";
+export { default as CheckboxTable_Deprecated } from "./components/CheckboxTable";
 export { default as Column } from "./components/Column";
 export { default as Cell } from "./components/Cell";
 export { default as HeaderCell } from "./components/HeaderCell";
