+++ packages/table_deprecated/stories/helpers/ResizableTableOnResizeDemo.tsx
